pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dimba-lib"
version = "0.1.0"
description = "DIMBA: Diffusion-based Mamba for non-autoregressive text generation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Faris Allafi", email = "faris.qais.allafi@gmail.com"}
]
keywords = ["diffusion", "mamba", "text-generation", "non-autoregressive", "language-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "mamba-ssm>=2.2.0",
    "transformers>=4.30.0",
    "datasets>=2.13.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
gpu = ["causal-conv1d>=1.4.0"]
eval = [
    "sacrebleu>=2.3.1",
    "rouge-score>=0.1.2",
    "nltk>=3.8.0",
]
tracking = [
    "wandb>=0.14.0",
    "tensorboard>=2.12.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "dimba-lib[gpu,eval,tracking,dev]",
]

[project.urls]
Repository = "https://github.com/devnull37/dimba-lib"
Documentation = "https://github.com/devnull37/dimba-lib#readme"
Issues = "https://github.com/devnull37/dimba-lib/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dimba"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/dimba --cov-report=term-missing"
